fix handleActivation service
