+++ lib/controller/challenges.js
@@ -44,9 +44,10 @@ app.controller('ChallengesController', function($scope, $rootScope, $routeParams
             }
 
             if (HandleActivationService.AreArguementsQueuedForProcessing()) {
+                console.log('We have some arguments for the handleActivation!q')
                 HandleActivationService.TestAndHandleArgs()
             } else {
-                alert('HELLO I am from the future! And I saw this bug!')
+                // 
             }
 
             if (routeWorld && !$scope.isWorldLocked(routeWorld)) {
