+++ lib/index.js
@@ -17,6 +17,7 @@ import './routes.js';
 import './service/email.js';
 import './service/content.js';
 import './service/social.js';
+import './service/HandleActivation.js';
 
 // Controllers
 import './controller/challenges.js';
